Refactored UI Components to UI Elements
This commit includes a broad refactor of the UIComponent and UIComponentHolder classes to UIElement and UIElementHolder respectively in the Screen package. The refactoring includes the renaming of packages, classes, and methods, for example, the change from UIComponentHolder::init to UIElementHolder::init. Additionally, the AbstractWidget, AbstractConfigOption, and StringOption classes have been updated according to the refactor. A new class, AbstractWidgetConverter has been introduced. The PLScreen implementation has also been adjusted.