Commit because i need to reinstall my PC Refactor code related to screen elements and add tooltip functionality
The primary operation in this commit involved enhancing the code of various screen elements, including the compilation of PLScreen and UIElementHolder. Moreover, the tooltip functionality has been added for better user interaction. Part of this effort also included reorganizing the placement of some functions and performing necessary file renaming for improved code readability and maintenance.