Add GPL license headers to source files
This commit adds General Public License (GPL) version 3 headers to all the source files in the project. This is to make it clear that the code is open source and redistributable under the terms and conditions of the GPL. It also clarifies the copyright ownership by Oliver Froberg (The Panda Oliver).