Update dependency versions and replace parse methods
The dependency versions of "fabric-api", "minecraft", "architectury", etc. are updated in the configuration files. The "ResourceLocation.parse()" method used in the networking scripts is replaced with the safer "ResourceLocation.tryParse()" method to avoid potential exceptions. Also, the way of creating "ResourceLocation" instances is updated by using the constructor directly, instead of deprecated "fromNamespaceAndPath" and "parse" methods.